# src/main/resources/specs/collector-manager-api-spec.yaml
openapi: 3.0.3
info:
  title: Recolnat
  description: |-
    Le réseau national des collections naturalistes 
    est une infrastructure de recherche française dont l'objectif est 
    de produire un corpus de données basé sur les collections d'histoire naturelle de France, 
    permettant de valoriser les recherches au service de l'étude de la géo- et de la biodiversité actuelle et passée.
    Liens utiles:
    - [RECOLNAT](https://fr.wikipedia.org/wiki/RECOLNAT)
  version: "1.0"
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@mnhn.fr
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: https://petstore3.swagger.io/api/v3
...
security:
  - collection-oidc: [ ] # <--- Use the same name as specified in securitySchemes*, here without scope
paths:
...
  /v1/public/articles/{id}:
    get:
      operationId: getPublicArticleById
      tags:
        - visibilite public
      summary: get Article By Id (for public acces)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
            format: uuid
            description: article identifier
            example:
              id: 386ff0b0-a6cc-11ec-b909-0242ac120002

      responses:
        '200':
          description: Successful retrieval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /v1/public/articles:
    get:
      operationId: getPublicArticles
      tags:
        - visibilite public
      summary: Retrieve all articles by pages  (for public acces)
      parameters:
        - in: query
          required: true
          name: page
          schema:
            type: integer
            format: int32
          description: page to access
        - in: query
          required: true
          name: size
          schema:
            type: integer
            format: int32
          description: size of elements per page
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResultPageResponse'
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /v1/public/specimens/{specimenId}:
    get:
      operationId: getPublicSpecimen
      tags:
        - visibilite public
      summary: get specimen from Id. Fetch specimen (for public acces)
      parameters:
        - name: specimenId
          required: true
          in: path
          schema:
            type: string
            format: uuid
            description: specimen identifier
          example:
            id: 123e4567-e89b-42d3-a456-556642440000
      responses:
        '200':
          description: return specimen details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSpecimen'
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /v1/public/collections/{id}:
    get:
      operationId: retrievePublicCollectionById
      tags:
        - visibilite public
      summary: retrieve collection by identifier (for public acces)
      parameters:
        - in: path
          required: true
          name: id
          schema:
            type: string
          description: collection identifier
      responses:
        '200':
          description: retrieve collection identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionPublic"
...
  /v1/public/programs:
    post:
      operationId: getPublicInstitutionPrograms
      tags:
        - visibilite public
      summary: Retrieve intitutions list for display name and logo (for public acces)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              nullable: false
              properties:
                codeinstitutions:
                  type: array
                  items:
                    type: string
                    format: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InstitutionsProgramResponse"
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /v1/public/statistics:
    get:
      tags:
        - visibilite public
      summary: get home page statistics (for public acces)
      operationId: getStatistic
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsResult'
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
    put:
      tags:
        - visibilite public
      operationId: UpdateStatistics
      summary: Update statistics counts
      responses:
        '204':
          description: Successful updated statistics operation
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /v1/public/validelastic:
    get:
      tags:
        - visibilite public
      summary: test connection elastic (ping) (for public acces)
      operationId: validElasticConnection
      responses:
        '200':
          description: Successful operation
        '403':
          description: Not found or Forbidden
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /v1/public/indexExist:
    get:
      tags:
        - visibilite public
      summary: test si l index elastic existe (for public acces)
      operationId: indexElasticExist
      parameters:
        - in: query
          required: true
          name: indexName
          schema:
            type: string
            description: nom de l index elastic
          example:
            name: recolnat_specimen
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  exist:
                    type: boolean
                example:
                  exist: true
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /v1/public/listInstitutions:
    post:
      operationId: getPublicListInstitutions
      tags:
        - visibilite public
      summary: Retrieve institutions from  uuid id list  (for public acces)
      requestBody:
        required: true
        content:
          application/json: # Media type
            schema: # Request payload
              type: array
              items:
                type: string
                format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstitutionListResponse'
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /v1/public/listCollections:
    post:
      operationId: getPublicListCollections
      tags:
        - visibilite public
      summary: Retrieve collections from  uuid id list  (for public acces)
      requestBody:
        required: true
        content:
          application/json: # Media type
            schema: # Request payload
              type: array
              items:
                type: string
                format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionListResponse'
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
components:
  securitySchemes:
    collection-oidc: # <--- Arbitrary name for the security scheme. Used to refer to it from elsewhere.
      type: openIdConnect
      openIdConnectUrl: http://sample.com/realms/recolnat-oauth2/.well-known/openid-configuration
  schemas:
...
    PublicSpecimen:
      description: Modèle représentant un spécimen sur la partie publique de l'application
      type: object
      nullable: false
      properties:
        id:
          type: string
          nullable: false
          example: f4b71ce1-9f8a-4dca-af89-4b61d6ec6884
        catalogNumber:
          type: string
          nullable: false
        collectionCode:
          type: string
        collectionName:
          type: string
          nullable: false
        collectionId:
          type: string
          format: uuid
          nullable: false
        institutionCode:
          type: string
          nullable: false
        institutionName:
          type: string
          nullable: false
        nominativeCollection:
          type: string
        recordNumber:
          type: string
        basisOfRecord:
          type: string
        individualCount:
          type: string
        sex:
          type: string
        lifeStage:
          type: string
        occurrenceRemarks:
          type: string
        legalStatus:
          type: string
        donor:
          type: string
        preparations:
          type: string
        identifications:
          type: array
          nullable: false
          items:
            $ref: '#/components/schemas/PublicIdentification'
        medias:
          type: array
          nullable: false
          items:
            $ref: '#/components/schemas/PublicMedia'
        collectionEvent:
          $ref: '#/components/schemas/PublicCollectionEvent'
        other:
          $ref: '#/components/schemas/PublicOther'
        geologicalContext:
          $ref: '#/components/schemas/PublicGeologicalContext'
        literatures:
          type: array
          nullable: false
          items:
            $ref: '#/components/schemas/PublicLiterature'
    PublicGeologicalContext:
      type: object
      properties:
        earliestAgeOrLowestStage:
          type: string
        latestAgeOrHighestStage:
          type: string
        earliestEpochOrLowestSeries:
          type: string
        latestEpochOrHighestSeries:
          type: string
        earliestPeriodOrLowestSystem:
          type: string
        latestPeriodOrHighestSystem:
          type: string
        earliestEraOrLowestErathem:
          type: string
        latestEraOrHighestErathem:
          type: string
        earliestEonOrLowestEonothem:
          type: string
        latestEonOrHighestEonothem:
          type: string
        lowestBiostratigraphicZone:
          type: string
        highestBiostratigraphicZone:
          type: string
        verbatimEpoch:
          type: string
        ageAbsolute:
          type: string
        range:
          type: string
        group:
          type: string
        formation:
          type: string
        member:
          type: string
        bed:
          type: string
        otherLithostratigraphicTerms:
          type: string
    PublicOther:
      type: object
      properties:
        linkGerBank:
          type: string
        linkBold:
          type: string
        linkOther:
          type: string
        financialAid:
          type: string
        computerizationProgram:
          type: string
    PublicCollectionEvent:
      type: object
      properties:
        eventDate:
          type: string
        recordedBy:
          type: string
        fieldNumber:
          type: string
        interpretedDate:
          type: boolean
        fieldNotes:
          type: string
        eventRemarks:
          type: string
        verbatimLocality:
          type: string
        sensitiveLocation:
          type: boolean
        location:
          $ref: '#/components/schemas/Location'
        minimumElevationInMeters:
          type: number
        maximumElevationInMeters:
          type: number
        interpretedAltitude:
          type: boolean
        minimumDepthInMeters:
          type: number
        maximumDepthInMeters:
          type: number
        interpretedDepth:
          type: boolean
        decimalLatitude:
          type: number
        decimalLongitude:
          type: number
        geodeticDatum:
          type: string
        georeferenceSources:
          type: string
        habitat:
          type: string
        locationRemarks:
          type: string
    PublicLiterature:
      type: object
      properties:
        id:
          type: string
          nullable: false
        authors:
          type: string
        citation:
          type: string
        identifier:
          type: string
        date:
          type: string
        title:
          type: string
        review:
          type: string
        volume:
          type: string
        number:
          type: string
        pages:
          type: string
        publisher:
          type: string
        publicationPlace:
          type: string
        editors:
          type: string
        bookTitle:
          type: string
        pageNumber:
          type: string
        language:
          type: string
        keywords:
          type: string
        description:
          type: string
        url:
          type: string
        remarks:
          type: string
...
    PublicIdentification:
      type: object
      nullable: false
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        currentDetermination:
          type: boolean
          nullable: false
        identifiedByID:
          type: string
        dateIdentified:
          type: string
        typeStatus:
          type: string
        identificationRemarks:
          type: string
        verbatimIdentification:
          type: string
        taxons:
          type: array
          nullable: false
          items:
            $ref: '#/components/schemas/PublicTaxon'
    PublicMedia:
      type: object
      nullable: false
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        creator:
          type: string
        contributor:
          type: string
        license:
          type: string
        source:
          type: string
        description:
          type: string
        fileName:
          type: string
          nullable: false
        mediaUrl:
          type: string
          nullable: false
        isCover:
          type: boolean
          description: if this media is a cover or not
...
    CollectionPublic:
      description: Collection data response public
      type: object
      nullable: false
      properties:
        id:
          type: string
          nullable: false
        typeCollection:
          type: string
          nullable: false
          example: h
        institutionCode:
          type: string
          nullable: false
          example: MNHN
        collectionNameFr:
          type: string
          nullable: false
          example: botanique
        collectionNameEn:
          type: string
          nullable: false
          example: botanique
        collectionName:
          type: string
          nullable: false
          example: botanique
        institutionId:
          type: string
          format: int
          description: institution id
          example: 86360ea7-b055-485b-8f46-8b75e4c3c1af
...
    Location:
      type: object
      properties:
        municipality:
          type: string
        stateProvince:
          type: string
        county:
          type: string
        country:
          type: string
        continent:
          type: string
        countryCode:
          type: string
        waterBody:
          type: string
        islandGroup:
          type: string
        island:
          type: string
        locality:
          type: string
        locationRemarks:
          type: string
...
    InstitutionListResponse:
      description: list Institutions Response
      type: array
      nullable: false
      items:
        $ref: "#/components/schemas/Institution"
    CollectionListResponse:
      description: list collection Response
      type: array
      nullable: false
      items:
        $ref: "#/components/schemas/CollectionDetailPublic"
...
    InstitutionsProgramResponse:
      description: Institutions programs Response
      type: object
      nullable: false
      properties:
        id:
          type: integer
          format: int32
          description: id sous forme int de l institution pour keycloak
          example: 159
        institutionId:
          type: string
          format: uuid
          description: id sous forme UUID de l institution
          example: 50f4978a-da62-4fde-8f38-5003bd43ff64
        code:
          type: string
          nullable: false
          description: code , libelle court de l institution
          example: MNHN
        name:
          type: string
          nullable: false
          description: nom de l institution
          example: Muséum National d''Histoire Naturelle
        logoUrl:
          type: string
          example: https://mediaphoto.mnhn.fr/media/1652113411737lMgX2dOG5396jVeF
...
    ArticleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        author:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        creationDate:
          type: string
          format: date
        urlMedia:
          type: string
        media:
          type: string
          format: binary
        incrementId:
          type: integer
    ArticleResultPageResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        articles:
          type: array
          items:
            $ref: '#/components/schemas/ArticleResponse'
    StatisticsResult:
      type: object
      properties:
        countSpecimen:
          type: integer
          format: int32
        countTaxon:
          type: integer
          format: int32
        countInstitution:
          type: integer
          format: int32
...
    UnexpectedError:
      $ref: '#/components/schemas/Error'
    Forbidden:
      $ref: '#/components/schemas/Error'
    Unauthorized:
      $ref: '#/components/schemas/Error'
    Conflict:
      $ref: '#/components/schemas/Error'
    Error:
      type: object
      description: A common error response
      properties:
        errorCode:
          type: integer
          description: Error Code
        errorMessage:
          type: string
          description: Message
        errorLevel:
          type: string
          description: Level of error
          default: info
          enum:
            - info
            - warning
            - error
            - critical
        errorType:
          type: string
          description: Type of error
          enum:
            - functional
            - technical
        documentationUrl:
          type: string
          description: Url of the documenation
        tips:
          type: string
          description: Tips/help for adressing the error
...